Arquivos para serem instalados no ec2

+++ requirements/ec2.sh
@@ -1,4 +1,16 @@
 apt-get install git
 apt-get install python-pip
 apt-get install libtiff4-dev libjpeg8-dev zlib1g-dev libfreetype6-dev liblcms2-dev libwebp-dev tcl8.5-dev tk8.5-dev
-pip install virtualenvwrapper
+
+echo "Configure VirutalEnvWrapper? [y/n]"
+read n
+y=$(echo $n | tr -s '[:upper:]' '[:lower:]')
+if [[  "$n" = "y"  ]] ; then
+  pip install virtualenvwrapper
+  echo "export WORKON_HOME=~/Envs
+mkdir -p $WORKON_HOME
+source /usr/local/bin/virtualenvwrapper.sh" >> ~/.bashrc
+  export WORKON_HOME=~/Envs
+  mkdir -p $WORKON_HOME
+  source /usr/local/bin/virtualenvwrapper.sh
+fi
